Add Entity and type definition tags to class

+++ backend/src/main/java/com/cmpt373sedna/gitlabanalyzer/repository/RepositoryRepository.java
@@ -3,5 +3,5 @@ package com.cmpt373sedna.gitlabanalyzer.repository;
 import com.cmpt373sedna.gitlabanalyzer.model.Repository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RepositoryRepository extends CrudRepository<Repository, Long> {
+public interface RepositoryRepository extends CrudRepository<Repository, Integer> {
 }
